Restructured the files to tests/transceiver folder
